+++ vim/dotfiles/.config/nvim/lua/plugins/notify.lua
@@ -1,5 +1,6 @@
 return {
     "rcarriga/nvim-notify",
+    tag = "v3.13.2",
     config = function ()
         vim.notify = require("notify")
     end